requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
clip==1.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
filelock==3.11.0
fonttools==4.39.3
ftfy==6.1.1
gdown==4.7.1
hdbscan==0.8.29
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
packaging==23.0
pandas==2.0.0
parso==0.8.3
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4
stack-data==0.6.2
sympy==1.11.1
threadpoolctl==3.1.0
torch==2.0.0
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
zipp==3.15.0
